Refactored examples to avoid the multi-main declarations
